Support Swift 3 and latest SourceKitten
