+++ .gitignore
@@ -18,3 +18,4 @@ Carthage/Build
 .build
 dist
 SourceKittenDaemon.pkg
+Packages
