--- Cartfile
@@ -1,3 +1,0 @@
-github "izqui/taylor" "697446f"
-github "Carthage/Commandant" == 0.7.0
-github "jpsim/SourceKitten" == 0.11.0
