--- Cartfile.resolved
@@ -1,6 +1,0 @@
-github "antitypical/Result" "1.0.2"
-github "drmohundro/SWXMLHash" "2.1.0"
-github "alwaysrightinstitute/SwiftSockets" "v0.20.6"
-github "Carthage/Commandant" "0.7.0"
-github "jpsim/SourceKitten" "0.11.0"
-github "izqui/taylor" "697446f8b2fa21ff0cdb3a55ab89f6d7196d27a8"
