--- SourceKittenDaemonTests/Helpers/regexp.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-infix operator =~ {}
-func =~(string:String, regex:String) -> Bool {
-    return string.rangeOfString(regex, options: .RegularExpressionSearch) != nil
-}
