+++ Sources/SourceKittenDaemon/Extensions/Path+.swift
@@ -1,0 +1,17 @@
+import Foundation
+import Xcode
+
+extension Path {
+
+    func absoluteURL(forProject project: Project) -> URL? {
+        switch self {
+        case .absolute(let path):
+            return URL(fileURLWithPath: path).standardizedFileURL
+        case .relativeTo(.sourceRoot, let path):
+            return (project.srcRoot as NSURL).appendingPathComponent(path)?.standardizedFileURL
+        case .relativeTo(_, let path):
+            return URL(fileURLWithPath: path).standardizedFileURL
+        }
+    }
+
+}
