+++ Sources/sourcekittend/main.swift
@@ -1,0 +1,21 @@
+//
+//  main.swift
+//  SourceKittenDaemon
+//
+//  Created by Benedikt Terhechte on 12/11/15.
+//  Copyright © 2015 Benedikt Terhechte. All rights reserved.
+//
+
+import Foundation
+import Commandant
+import SourceKittenDaemon
+
+let registry = CommandRegistry<CommandError>()
+registry.register(StartCommand())
+
+let helpCommand = HelpCommand(registry: registry)
+registry.register(helpCommand)
+
+registry.main(defaultVerb: "help") { error in
+    fputs("\(error)\n", stderr)
+}
