--- SwiftCode/AppDelegate.swift
@@ -1,28 +1,0 @@
-//
-//  AppDelegate.swift
-//  SwiftCode
-//
-//  Created by Benedikt Terhechte on 03/12/15.
-//  Copyright © 2015 Benedikt Terhechte. All rights reserved.
-//
-
-import Cocoa
-
-@NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate {
-
-    @IBOutlet weak var window: NSWindow!
-    @IBOutlet var controller: EditorController!
-
-
-    func applicationDidFinishLaunching(aNotification: NSNotification) {
-        // Insert code here to initialize your application
-    }
-
-    func applicationWillTerminate(aNotification: NSNotification) {
-        self.controller.terminate()
-    }
-
-
-}
-
