+++ old/Tests.bak/Fixtures/CompleteMethodFromFrameworkFixture.swift
