+++ old/Tests.bak/Helpers/regexp.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+infix operator =~
+func =~(string:String, regex:String) -> Bool {
+    return string.range(of: regex, options: .regularExpression) != nil
+}
