--- script/extract-tool
@@ -1,6 +1,0 @@
-#!/bin/bash
-#
-# Extracts the sourcekitten CLI tool from its application bundle. Meant to be run
-# as part of an Xcode Run Script build phase.
-
-cp -v "${BUILT_PRODUCTS_DIR}/${EXECUTABLE_PATH}" "${BUILT_PRODUCTS_DIR}/${EXECUTABLE_NAME}"
